Code updated

+++ js/main.js
@@ -5,7 +5,7 @@ let IsLoggedIn = false;
 
 // Login function
 function Login() {
-    fetch("./js/users.json").then(res => res.json())
+    fetch("../js/users.json").then(res => res.json())
         .then((data) => {
             data.forEach((result, index) => {
                 if (result.username === user.value) {
@@ -93,7 +93,7 @@ window.onload = function afterLogin() {
 
 if (showLogin) {
     if (Data.name === "Admin") {
-        document.getElementById('exploreBtn').style.display = 'none';
+        document.getElementById('exploreBtn').style.visibility = 'hidden';
     }
 }
 
@@ -147,6 +147,21 @@ function serviceDropdown() {
 }
 serviceDropdown()
 
+// Booking successfull popups
+function bookingSuccess() {
+    const getUrl = window.location.pathname.split("AutoMob/")
+
+    if (getUrl[1] === "booking-success.html") {
+        setTimeout(() => {
+            prompt('On scale of 1-10, howlikely are you to recommend our website to your friends & family?', 100);
+            alert(`Thank you for your feedback`);
+            location.assign("index.html")
+        }, 1000)
+    }
+}
+
+bookingSuccess()
+
 // Booking submit data
 const nameFull = document.getElementById("nameFull");
 const emailId = document.getElementById("emailId");
@@ -158,53 +173,43 @@ const fuelType = document.getElementById("fuelType");
 const appointmentDate = document.getElementById("appointmentDate");
 const address = document.getElementById("address");
 const formElem = document.getElementById("formElem");
-
-formElem.addEventListener("submit", (e) => {
-    e.preventDefault()
-    let formData = new FormData();
-    formData.append('name', nameFull.value)
-    formData.append('email', emailId.value)
-    formData.append('phone', phone.value)
-    formData.append('service', selectService.value)
-    formData.append('make', make.value)
-    formData.append('model', model.value)
-    // formData.append('fuelType', fuelType.value)
-    formData.append('appointment', appointmentDate.value)
-    formData.append('address', address.value)
-
-    var bookings = [];
-    var obj = {};
-
-    for (let [name, value] of formData) {
-        obj[name] = value;
-
-        bookings.push(obj)
-    }
-
-    const data = JSON.stringify(bookings)
-
-    localStorage.setItem("bookings", data)
-
-    var retrievedPerson = JSON.parse(localStorage.getItem('bookings'))
-
-    console.log(retrievedPerson)
-
-    localStorage.clear();
-})
-
-
-// Booking successfull popups
-function bookingSuccess() {
-    const getUrl = window.location.pathname.split("AutoMob/")
-    if (getUrl[1] === "booking-success.html") {
-        setTimeout(() => {
-            prompt('On scale of 1-10, howlikely are you to recommend our website to your friends & family?', 100);
-            alert(`Thank you for your feedback`);
-            location.assign("index.html")
-        }, 1000)
-    }
+    
+if(formElem){
+    formElem.addEventListener('submit', (e)=>{
+        e.preventDefault()
+        let formData = new FormData();
+        formData.append('name', nameFull.value)
+        formData.append('email', emailId.value)
+        formData.append('phone', phone.value)
+        formData.append('service', selectService.value)
+        formData.append('make', make.value)
+        formData.append('model', model.value)
+        // formData.append('fuelType', fuelType.value)
+        formData.append('appointment', appointmentDate.value)
+        formData.append('address', address.value)
+    
+        var bookings = [];
+        var obj = {};
+    
+        for(let [name, value] of formData) {
+            obj[name] = value;
+            
+            bookings.push(obj)
+        }
+    
+        const data = JSON.stringify(bookings)
+    
+        localStorage.setItem("bookings", data)
+    
+        var retrievedPerson = JSON.parse(localStorage.getItem('bookings'));
+    
+        alert("Service booked successfully!!")
+    
+        localStorage.clear();
+    
+        location.assign('booking-success.html')
+    })
 }
-bookingSuccess()
 
 //Logout
 const logout = document.getElementById("logout");
